Prepare for new release 0.1.0
Large updates to docs
