+++ src/clone.rs
@@ -1,46 +1,14 @@
-use crate::{GitCredentials, GitRepo, GitRepoCloner};
+use crate::{GitCredentials, GitRepo};
 
-use color_eyre::eyre::Result;
-use git2::Cred;
-use git_url_parse::GitUrl;
-use log::{debug, info};
 use std::path::Path;
 use std::process::{Command, Stdio};
 
-impl From<GitRepo> for GitRepoCloner {
-    fn from(repo: GitRepo) -> GitRepoCloner {
-        GitRepoCloner {
-            url: repo.url,
-            credentials: repo.credentials,
-            branch: repo.branch,
-            path: repo.path,
-        }
-    }
-}
+use color_eyre::eyre::Result;
+use log::{debug, info};
 
-impl GitRepoCloner {
-    pub fn new<S: AsRef<str>>(url: S) -> Result<GitRepoCloner> {
-        Ok(GitRepoCloner {
-            url: GitUrl::parse(url.as_ref()).expect("url failed to parse as GitUrl"),
-            credentials: None,
-            // This is only used by clone()
-            branch: None,
-            path: None,
-        })
-    }
-
-    pub fn with_credentials(mut self, creds: GitCredentials) -> Self {
-        self.credentials = Some(creds);
-        self
-    }
-
-    pub fn with_branch<S: ToString>(mut self, branch: S) -> Self {
-        self.branch = Some(branch.to_string());
-        self
-    }
-
-    // TODO: Change return to Result<GitRepo>
-    pub fn git_clone<P: AsRef<Path>>(&self, target: P) -> Result<git2::Repository> {
+impl GitRepo {
+    // TODO: Can we make this mut self?
+    pub fn git_clone<P: AsRef<Path>>(&self, target: P) -> Result<GitRepo> {
         let cb = self.build_git2_remotecallback();
 
         let mut builder = git2::build::RepoBuilder::new();
@@ -58,18 +26,22 @@ impl GitRepoCloner {
             Err(e) => panic!("failed to clone: {}", e),
         };
 
-        Ok(repo)
+        // Ensure we don't lose the credentials while updating
+        let mut git_repo: GitRepo = repo.into();
+        git_repo = git_repo.with_credentials(self.credentials.clone());
+
+        Ok(git_repo)
     }
 
-    // TODO: Change return to Result<GitRepo>
-    pub fn git_clone_shallow<P: AsRef<Path>>(&self, target: P) -> Result<git2::Repository> {
+    // TODO: Can we make this mut self?
+    pub fn git_clone_shallow<P: AsRef<Path>>(&self, target: P) -> Result<GitRepo> {
         let repo = if let Some(cred) = self.credentials.clone() {
             match cred {
                 crate::GitCredentials::SshKey {
                     username,
-                    public_key: _,
+                    public_key,
                     private_key,
-                    passphrase: _,
+                    passphrase,
                 } => {
                     let mut parsed_uri = self.url.trim_auth();
                     parsed_uri.user = Some(username.to_string());
@@ -84,6 +56,7 @@ impl GitRepoCloner {
                         .arg(format!(
                             "core.sshcommand=ssh -i {privkey_path}",
                             privkey_path = private_key
+                                .clone()
                                 .into_os_string()
                                 .into_string()
                                 .expect("Couldn't convert path to string")
@@ -99,8 +72,19 @@ impl GitRepoCloner {
 
                     debug!("Clone output: {:?}", clone_out);
 
-                    git2::Repository::open(target.as_ref())
-                        .expect("Failed to open shallow clone dir")
+                    // Re-create the GitCredentials
+                    let creds = GitCredentials::SshKey {
+                        username,
+                        public_key,
+                        private_key,
+                        passphrase,
+                    };
+
+                    GitRepo::open(target.as_ref().to_path_buf(), None, None)
+                        .expect(
+                            format!("Failed to open shallow clone dir: {:?}", clone_out).as_str(),
+                        )
+                        .with_credentials(Some(creds))
                 }
                 crate::GitCredentials::UserPassPlaintext { username, password } => {
                     let mut cli_remote_url = self.url.clone();
@@ -119,9 +103,15 @@ impl GitRepoCloner {
                         .expect("Failed to run git clone");
 
                     let clone_out = shell_clone_command.stdout.expect("Failed to open stdout");
-                    git2::Repository::open(target.as_ref()).expect(
-                        format!("Failed to open shallow clone dir: {:?}", clone_out).as_str(),
-                    )
+
+                    // Re-create the GitCredentials
+                    let creds = GitCredentials::UserPassPlaintext { username, password };
+
+                    GitRepo::open(target.as_ref().to_path_buf(), None, None)
+                        .expect(
+                            format!("Failed to open shallow clone dir: {:?}", clone_out).as_str(),
+                        )
+                        .with_credentials(Some(creds))
                 }
             }
         } else {
@@ -146,80 +136,10 @@ impl GitRepoCloner {
                 .expect("Failed to wait for output")
                 .stdout;
 
-            git2::Repository::open(target.as_ref())
+            GitRepo::open(target.as_ref().to_path_buf(), None, None)
                 .expect(format!("Failed to open shallow clone dir: {:?}", clone_out).as_str())
         };
 
         Ok(repo)
-    }
-
-    // FIXME: This is a copy
-    pub fn build_git2_remotecallback(&self) -> git2::RemoteCallbacks {
-        if let Some(cred) = self.credentials.clone() {
-            debug!("Before building callback: {:?}", &cred);
-
-            match cred {
-                GitCredentials::SshKey {
-                    username,
-                    public_key,
-                    private_key,
-                    passphrase,
-                } => {
-                    let mut cb = git2::RemoteCallbacks::new();
-                    let privkey_path = std::path::PathBuf::from(private_key);
-
-                    cb.credentials(
-                        move |_, _, _| match (public_key.clone(), passphrase.clone()) {
-                            (None, None) => {
-                                Ok(Cred::ssh_key(&username, None, privkey_path.as_path(), None)
-                                    .expect("Could not create credentials object for ssh key"))
-                            }
-                            (None, Some(pp)) => Ok(Cred::ssh_key(
-                                &username,
-                                None,
-                                privkey_path.as_path(),
-                                Some(pp.as_ref()),
-                            )
-                            .expect("Could not create credentials object for ssh key")),
-                            (Some(pk), None) => {
-                                let pubkey_path = std::path::PathBuf::from(pk);
-
-                                Ok(Cred::ssh_key(
-                                    &username,
-                                    Some(pubkey_path.as_path()),
-                                    privkey_path.as_path(),
-                                    None,
-                                )
-                                .expect("Could not create credentials object for ssh key"))
-                            }
-                            (Some(pk), Some(pp)) => {
-                                let pubkey_path = std::path::PathBuf::from(pk);
-
-                                Ok(Cred::ssh_key(
-                                    &username,
-                                    Some(pubkey_path.as_path()),
-                                    privkey_path.as_path(),
-                                    Some(pp.as_ref()),
-                                )
-                                .expect("Could not create credentials object for ssh key"))
-                            }
-                        },
-                    );
-
-                    cb
-                }
-                GitCredentials::UserPassPlaintext { username, password } => {
-                    let mut cb = git2::RemoteCallbacks::new();
-                    cb.credentials(move |_, _, _| {
-                        Cred::userpass_plaintext(username.as_str(), password.as_str())
-                    });
-
-                    cb
-                }
-            }
-        } else {
-            // No credentials. Repo is public
-            git2::RemoteCallbacks::new()
-        }
     }
 }
